+++ tests/net/ptp/clock/src/main.c
@@ -218,7 +218,7 @@ static int my_ptp_clock_adjust(const struct device *dev, int increment)
 	return 0;
 }
 
-static int my_ptp_clock_rate_adjust(const struct device *dev, float ratio)
+static int my_ptp_clock_rate_adjust(const struct device *dev, double ratio)
 {
 	return 0;
 }
